#47: Support for global tools (#113)
