+++ cli/src/main/java/com/devonfw/tools/ide/tool/ToolCommandlet.java
@@ -4,7 +4,6 @@ import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
-import java.nio.file.StandardOpenOption;
 import java.util.List;
 import java.util.Set;
 import java.util.stream.Stream;
@@ -16,14 +15,11 @@ import com.devonfw.tools.ide.context.IdeContext;
 import com.devonfw.tools.ide.environment.EnvironmentVariables;
 import com.devonfw.tools.ide.environment.EnvironmentVariablesType;
 import com.devonfw.tools.ide.io.FileAccess;
-import com.devonfw.tools.ide.io.FileCopyMode;
 import com.devonfw.tools.ide.io.TarCompression;
-import com.devonfw.tools.ide.log.IdeLogLevel;
 import com.devonfw.tools.ide.os.MacOsHelper;
 import com.devonfw.tools.ide.process.ProcessContext;
 import com.devonfw.tools.ide.process.ProcessErrorHandling;
 import com.devonfw.tools.ide.property.StringListProperty;
-import com.devonfw.tools.ide.repo.ToolRepository;
 import com.devonfw.tools.ide.util.FilenameUtil;
 import com.devonfw.tools.ide.version.VersionIdentifier;
 
@@ -68,6 +64,15 @@ public abstract class ToolCommandlet extends Commandlet implements Tags {
     return this.tool;
   }
 
+  /**
+   *
+   * @return the name of the binary
+   */
+  protected String getBinaryName() {
+
+    return this.tool;
+  }
+
   @Override
   public final Set<String> getTags() {
 
@@ -103,11 +108,6 @@ public abstract class ToolCommandlet extends Commandlet implements Tags {
     pc.run();
   }
 
-  protected String getBinaryName() {
-
-    return this.tool;
-  }
-
   /**
    * @return the {@link EnvironmentVariables#getToolEdition(String) tool edition}.
    */
@@ -141,63 +141,11 @@ public abstract class ToolCommandlet extends Commandlet implements Tags {
   }
 
   /**
-   * @return the {@link Path} where the tool is located (installed).
-   */
-  public Path getToolPath() {
-
-    return this.context.getSoftwarePath().resolve(getName());
-  }
-
-  /**
-   * @return the {@link Path} where the executables of the tool can be found. Typically a "bin" folder inside
-   *         {@link #getToolPath() tool path}.
-   */
-  public Path getToolBinPath() {
-
-    Path toolPath = getToolPath();
-    Path binPath = this.context.getFileAccess().findFirst(toolPath, path -> path.getFileName().toString().equals("bin"),
-        false);
-    if ((binPath != null) && Files.isDirectory(binPath)) {
-      return binPath;
-    }
-    return toolPath;
-  }
-
-  /**
    * @return the {@link EnvironmentVariables#getToolVersion(String) tool version}.
    */
   public VersionIdentifier getConfiguredVersion() {
 
     return this.context.getVariables().getToolVersion(getName());
-  }
-
-  /**
-   * @return the currently installed {@link VersionIdentifier version} of this tool or {@code null} if not installed.
-   */
-  public VersionIdentifier getInstalledVersion() {
-
-    Path toolPath = getToolPath();
-    if (!Files.isDirectory(toolPath)) {
-      this.context.trace("Tool {} not installed in {}", getName(), toolPath);
-      return null;
-    }
-    Path toolVersionFile = toolPath.resolve(IdeContext.FILE_SOFTWARE_VERSION);
-    if (!Files.exists(toolVersionFile)) {
-      Path legacyToolVersionFile = toolPath.resolve(IdeContext.FILE_LEGACY_SOFTWARE_VERSION);
-      if (Files.exists(legacyToolVersionFile)) {
-        toolVersionFile = legacyToolVersionFile;
-      } else {
-        this.context.warning("Tool {} is missing version file in {}", getName(), toolVersionFile);
-        return null;
-      }
-    }
-    try {
-      String version = Files.readString(toolVersionFile).trim();
-      return VersionIdentifier.of(version);
-    } catch (IOException e) {
-      throw new IllegalStateException("Failed to read file " + toolVersionFile, e);
-    }
-
   }
 
   /**
@@ -230,181 +178,14 @@ public abstract class ToolCommandlet extends Commandlet implements Tags {
    * @return {@code true} if the tool was newly installed, {@code false} if the tool was already installed before and
    *         nothing has changed.
    */
-  protected boolean doInstall(boolean silent) {
-
-    VersionIdentifier configuredVersion = getConfiguredVersion();
-    // install configured version of our tool in the software repository if not already installed
-    ToolInstallation installation = installInRepo(configuredVersion);
-
-    // check if we already have this version installed (linked) locally in IDE_HOME/software
-    VersionIdentifier installedVersion = getInstalledVersion();
-    VersionIdentifier resolvedVersion = installation.resolvedVersion();
-    if (isInstalledVersion(resolvedVersion, installedVersion, silent)) {
-      return false;
-    }
-    // we need to link the version or update the link.
-    Path toolPath = getToolPath();
-    FileAccess fileAccess = this.context.getFileAccess();
-    if (Files.exists(toolPath)) {
-      fileAccess.backup(toolPath);
-    }
-    fileAccess.symlink(installation.linkDir(), toolPath);
-    this.context.getPath().setPath(this.tool, installation.binDir());
-    if (installedVersion == null) {
-      this.context.success("Successfully installed {} in version {}", this.tool, resolvedVersion);
-    } else {
-      this.context.success("Successfully installed {} in version {} replacing previous version {}", this.tool,
-          resolvedVersion, installedVersion);
-    }
-    postInstall();
-    return true;
-  }
+  protected abstract boolean doInstall(boolean silent);
 
   /**
-   * This method is called after the tool has been newly installed or updated to a new version. Override it to add
-   * custom post intallation logic.
+   * This method is called after the tool has been newly installed or updated to a new version.
    */
   protected void postInstall() {
 
     // nothing to do by default
-  }
-
-  /**
-   * Performs the installation of the {@link #getName() tool} managed by this {@link Commandlet} only in the central
-   * software repository without touching the IDE installation.
-   *
-   * @param version the {@link VersionIdentifier} requested to be installed. May also be a
-   *        {@link VersionIdentifier#isPattern() version pattern}.
-   * @return the {@link ToolInstallation} in the central software repository matching the given {@code version}.
-   */
-  public ToolInstallation installInRepo(VersionIdentifier version) {
-
-    return installInRepo(version, getEdition());
-  }
-
-  /**
-   * Performs the installation of the {@link #getName() tool} managed by this {@link Commandlet} only in the central
-   * software repository without touching the IDE installation.
-   *
-   * @param version the {@link VersionIdentifier} requested to be installed. May also be a
-   *        {@link VersionIdentifier#isPattern() version pattern}.
-   * @param edition the specific edition to install.
-   * @return the {@link ToolInstallation} in the central software repository matching the given {@code version}.
-   */
-  public ToolInstallation installInRepo(VersionIdentifier version, String edition) {
-
-    return installInRepo(version, edition, this.context.getDefaultToolRepository());
-  }
-
-  /**
-   * Performs the installation of the {@link #getName() tool} managed by this {@link Commandlet} only in the central
-   * software repository without touching the IDE installation.
-   *
-   * @param version the {@link VersionIdentifier} requested to be installed. May also be a
-   *        {@link VersionIdentifier#isPattern() version pattern}.
-   * @param edition the specific edition to install.
-   * @param toolRepository the {@link ToolRepository} to use.
-   * @return the {@link ToolInstallation} in the central software repository matching the given {@code version}.
-   */
-  public ToolInstallation installInRepo(VersionIdentifier version, String edition, ToolRepository toolRepository) {
-
-    VersionIdentifier resolvedVersion = toolRepository.resolveVersion(this.tool, edition, version);
-    Path toolPath = this.context.getSoftwareRepositoryPath().resolve(toolRepository.getId()).resolve(this.tool)
-        .resolve(edition).resolve(resolvedVersion.toString());
-    Path toolVersionFile = toolPath.resolve(IdeContext.FILE_SOFTWARE_VERSION);
-    FileAccess fileAccess = this.context.getFileAccess();
-    if (Files.isDirectory(toolPath)) {
-      if (Files.exists(toolVersionFile)) {
-        this.context.debug("Version {} of tool {} is already installed at {}", resolvedVersion,
-            getToolWithEdition(this.tool, edition), toolPath);
-        return createToolInstallation(toolPath, resolvedVersion, toolVersionFile);
-      }
-      this.context.warning("Deleting corrupted installation at {}", toolPath);
-      fileAccess.delete(toolPath);
-    }
-    Path target = toolRepository.download(this.tool, edition, resolvedVersion);
-    fileAccess.mkdirs(toolPath.getParent());
-    extract(target, toolPath);
-    try {
-      Files.writeString(toolVersionFile, resolvedVersion.toString(), StandardOpenOption.CREATE_NEW);
-    } catch (IOException e) {
-      throw new IllegalStateException("Failed to write version file " + toolVersionFile, e);
-    }
-    return createToolInstallation(toolPath, resolvedVersion, toolVersionFile);
-  }
-
-  private ToolInstallation createToolInstallation(Path rootDir, VersionIdentifier resolvedVersion,
-      Path toolVersionFile) {
-
-    Path linkDir = getMacOsHelper().findLinkDir(rootDir);
-    Path binDir = linkDir;
-    Path binFolder = binDir.resolve(IdeContext.FOLDER_BIN);
-    if (Files.isDirectory(binFolder)) {
-      binDir = binFolder;
-    }
-    if (linkDir != rootDir) {
-      assert (!linkDir.equals(rootDir));
-      this.context.getFileAccess().copy(toolVersionFile, linkDir, FileCopyMode.COPY_FILE_OVERRIDE);
-    }
-    return new ToolInstallation(rootDir, linkDir, binDir, resolvedVersion);
-  }
-
-  private MacOsHelper getMacOsHelper() {
-
-    if (this.macOsHelper == null) {
-      this.macOsHelper = new MacOsHelper(this.context);
-    }
-    return this.macOsHelper;
-  }
-
-  /**
-   * @return {@code true} to extract (unpack) the downloaded binary file, {@code false} otherwise.
-   */
-  protected boolean isExtract() {
-
-    return true;
-  }
-
-  /**
-   * @return the currently installed tool version or {@code null} if not found (tool not installed).
-   */
-  protected String getInstalledToolVersion() {
-
-    Path toolPath = getToolPath();
-    if (!Files.isDirectory(toolPath)) {
-      this.context.debug("Tool {} not installed in {}", getName(), toolPath);
-      return null;
-    }
-    Path toolVersionFile = toolPath.resolve(IdeContext.FILE_SOFTWARE_VERSION);
-    if (!Files.exists(toolVersionFile)) {
-      Path legacyToolVersionFile = toolPath.resolve(IdeContext.FILE_LEGACY_SOFTWARE_VERSION);
-      if (Files.exists(legacyToolVersionFile)) {
-        toolVersionFile = legacyToolVersionFile;
-      } else {
-        this.context.warning("Tool {} is missing version file in {}", getName(), toolVersionFile);
-        return null;
-      }
-    }
-    try {
-      return Files.readString(toolVersionFile).trim();
-    } catch (IOException e) {
-      throw new IllegalStateException("Failed to read file " + toolVersionFile, e);
-    }
-  }
-
-  private boolean isInstalledVersion(VersionIdentifier expectedVersion, VersionIdentifier installedVersion,
-      boolean silent) {
-
-    if (expectedVersion.equals(installedVersion)) {
-      IdeLogLevel level = IdeLogLevel.INFO;
-      if (silent) {
-        level = IdeLogLevel.DEBUG;
-      }
-      this.context.level(level).log("Version {} of tool {} is already installed", installedVersion,
-          getToolWithEdition());
-      return true;
-    }
-    return false;
   }
 
   /**
@@ -439,9 +220,9 @@ public abstract class ToolCommandlet extends Commandlet implements Tags {
    */
   protected void extract(Path file, Path targetDir) {
 
-    Path tmpDir = this.context.getFileAccess().createTempDir("extract-" + file.getFileName());
     FileAccess fileAccess = this.context.getFileAccess();
     if (isExtract()) {
+      Path tmpDir = this.context.getFileAccess().createTempDir("extract-" + file.getFileName());
       this.context.trace("Trying to extract the downloaded file {} to {} and move it to {}.", file, tmpDir, targetDir);
       String extension = FilenameUtil.getExtension(file.getFileName().toString());
       this.context.trace("Determined file extension {}", extension);
@@ -491,6 +272,57 @@ public abstract class ToolCommandlet extends Commandlet implements Tags {
     } else {
       this.context.trace("Extraction is disabled for '{}' hence just moving the downloaded file {}.", getName(), file);
       fileAccess.move(file, targetDir);
+    }
+  }
+
+  /**
+   * @return {@code true} to extract (unpack) the downloaded binary file, {@code false} otherwise.
+   */
+  protected boolean isExtract() {
+
+    return true;
+  }
+
+  protected MacOsHelper getMacOsHelper() {
+
+    if (this.macOsHelper == null) {
+      this.macOsHelper = new MacOsHelper(this.context);
+    }
+    return this.macOsHelper;
+  }
+
+  /**
+   * @return the currently installed {@link VersionIdentifier version} of this tool or {@code null} if not installed.
+   */
+  public VersionIdentifier getInstalledVersion() {
+
+    return getInstalledVersion(this.context.getSoftwarePath().resolve(getName()));
+  }
+
+  /**
+   * @return the currently installed {@link VersionIdentifier version} of this tool or {@code null} if not installed.
+   */
+  protected VersionIdentifier getInstalledVersion(Path toolPath) {
+
+    if (!Files.isDirectory(toolPath)) {
+      this.context.debug("Tool {} not installed in {}", getName(), toolPath);
+      return null;
+    }
+    Path toolVersionFile = toolPath.resolve(IdeContext.FILE_SOFTWARE_VERSION);
+    if (!Files.exists(toolVersionFile)) {
+      Path legacyToolVersionFile = toolPath.resolve(IdeContext.FILE_LEGACY_SOFTWARE_VERSION);
+      if (Files.exists(legacyToolVersionFile)) {
+        toolVersionFile = legacyToolVersionFile;
+      } else {
+        this.context.warning("Tool {} is missing version file in {}", getName(), toolVersionFile);
+        return null;
+      }
+    }
+    try {
+      String version = Files.readString(toolVersionFile).trim();
+      return VersionIdentifier.of(version);
+    } catch (IOException e) {
+      throw new IllegalStateException("Failed to read file " + toolVersionFile, e);
     }
   }
