+++ cli/src/main/java/com/devonfw/tools/ide/tool/gradle/Gradle.java
@@ -3,13 +3,14 @@ package com.devonfw.tools.ide.tool.gradle;
 import java.util.Set;
 
 import com.devonfw.tools.ide.context.IdeContext;
+import com.devonfw.tools.ide.tool.LocalToolCommandlet;
 import com.devonfw.tools.ide.tool.ToolCommandlet;
 import com.devonfw.tools.ide.tool.java.Java;
 
 /**
  * {@link ToolCommandlet} for <a href="https://gradle.org/">gradle</a>.
  */
-public class Gradle extends ToolCommandlet {
+public class Gradle extends LocalToolCommandlet {
 
   /**
    * The constructor.
