+++ cli/src/main/java/com/devonfw/tools/ide/tool/helm/Helm.java
@@ -1,6 +1,7 @@
 package com.devonfw.tools.ide.tool.helm;
 
 import com.devonfw.tools.ide.context.IdeContext;
+import com.devonfw.tools.ide.tool.LocalToolCommandlet;
 import com.devonfw.tools.ide.tool.ToolCommandlet;
 
 import java.util.Set;
@@ -8,7 +9,7 @@ import java.util.Set;
 /**
  * {@link ToolCommandlet} for <a href="https://helm.sh/">Helm</a>, the package manager for Kubernetes.
  */
-public class Helm extends ToolCommandlet {
+public class Helm extends LocalToolCommandlet {
   /**
    * The constructor.
    *
