+++ cli/src/main/java/com/devonfw/tools/ide/tool/java/Java.java
@@ -3,12 +3,13 @@ package com.devonfw.tools.ide.tool.java;
 import java.util.Set;
 
 import com.devonfw.tools.ide.context.IdeContext;
+import com.devonfw.tools.ide.tool.LocalToolCommandlet;
 import com.devonfw.tools.ide.tool.ToolCommandlet;
 
 /**
  * {@link ToolCommandlet} for Java (Java Virtual Machine and Java Development Kit).
  */
-public class Java extends ToolCommandlet {
+public class Java extends LocalToolCommandlet {
 
   /**
    * The constructor.
