+++ cli/src/main/java/com/devonfw/tools/ide/tool/kotlinc/KotlincNative.java
@@ -3,12 +3,13 @@ package com.devonfw.tools.ide.tool.kotlinc;
 import java.util.Set;
 
 import com.devonfw.tools.ide.context.IdeContext;
+import com.devonfw.tools.ide.tool.LocalToolCommandlet;
 import com.devonfw.tools.ide.tool.ToolCommandlet;
 
 /**
  * {@link ToolCommandlet} for <a href="https://kotlinlang.org/docs/native-overview.html">Kotlin Native</a> (kotlincnative).
  */
-public class KotlincNative extends ToolCommandlet {
+public class KotlincNative extends LocalToolCommandlet {
 
   /**
    * The constructor.
