+++ cli/src/main/java/com/devonfw/tools/ide/tool/mvn/Mvn.java
@@ -3,13 +3,14 @@ package com.devonfw.tools.ide.tool.mvn;
 import java.util.Set;
 
 import com.devonfw.tools.ide.context.IdeContext;
+import com.devonfw.tools.ide.tool.LocalToolCommandlet;
 import com.devonfw.tools.ide.tool.ToolCommandlet;
 import com.devonfw.tools.ide.tool.java.Java;
 
 /**
  * {@link ToolCommandlet} for <a href="https://maven.apache.org/">maven</a>.
  */
-public class Mvn extends ToolCommandlet {
+public class Mvn extends LocalToolCommandlet {
 
   /**
    * The constructor.
