+++ cli/src/main/java/com/devonfw/tools/ide/tool/node/Node.java
@@ -3,12 +3,13 @@ package com.devonfw.tools.ide.tool.node;
 import java.util.Set;
 
 import com.devonfw.tools.ide.context.IdeContext;
+import com.devonfw.tools.ide.tool.LocalToolCommandlet;
 import com.devonfw.tools.ide.tool.ToolCommandlet;
 
 /**
  * {@link ToolCommandlet} for <a href="https://nodejs.org/">node</a>.
  */
-public class Node extends ToolCommandlet {
+public class Node extends LocalToolCommandlet {
 
   /**
    * The constructor.
