+++ cli/src/main/java/com/devonfw/tools/ide/tool/oc/Oc.java
@@ -1,6 +1,7 @@
 package com.devonfw.tools.ide.tool.oc;
 
 import com.devonfw.tools.ide.context.IdeContext;
+import com.devonfw.tools.ide.tool.LocalToolCommandlet;
 import com.devonfw.tools.ide.tool.ToolCommandlet;
 
 import java.util.Set;
@@ -8,7 +9,7 @@ import java.util.Set;
 /**
  * {@link ToolCommandlet} for <a href="https://docs.openshift.com/">Openshift CLI</a>.
  */
-public class Oc extends ToolCommandlet {
+public class Oc extends LocalToolCommandlet {
 
   /**
    * The constructor.
