+++ cli/src/main/java/com/devonfw/tools/ide/tool/quarkus/Quarkus.java
@@ -1,6 +1,7 @@
 package com.devonfw.tools.ide.tool.quarkus;
 
 import com.devonfw.tools.ide.context.IdeContext;
+import com.devonfw.tools.ide.tool.LocalToolCommandlet;
 import com.devonfw.tools.ide.tool.ToolCommandlet;
 
 import java.util.Set;
@@ -8,7 +9,7 @@ import java.util.Set;
 /**
  * {@link ToolCommandlet} for <a href="https://quarkus.io/">Quarkus</a>.
  */
-public class Quarkus extends ToolCommandlet {
+public class Quarkus extends LocalToolCommandlet {
   /**
    * The constructor
    *
