+++ cli/src/main/java/com/devonfw/tools/ide/tool/vscode/Vscode.java
@@ -3,12 +3,13 @@ package com.devonfw.tools.ide.tool.vscode;
 import java.util.Set;
 
 import com.devonfw.tools.ide.context.IdeContext;
+import com.devonfw.tools.ide.tool.LocalToolCommandlet;
 import com.devonfw.tools.ide.tool.ToolCommandlet;
 
 /**
  * {@link ToolCommandlet} for <a href="https://code.visualstudio.com/">vscode</a>.
  */
-public class Vscode extends ToolCommandlet {
+public class Vscode extends LocalToolCommandlet {
 
   /**
    * The constructor.
@@ -22,6 +23,7 @@ public class Vscode extends ToolCommandlet {
 
   @Override
   protected String getBinaryName() {
+
     return "code";
   }
